migrate to Rust 2021 edition and Update the crates to the latest version

+++ machine/src/standard_vm/aarch64/mod.rs
@@ -440,7 +440,7 @@ impl MachineOps for StdMachine {
                         .vm_fd
                         .as_ref()
                         .unwrap()
-                        .create_vcpu(vcpu_id)?,
+                        .create_vcpu(vcpu_id as u64)?,
                 ));
             }
             fds
